refactor: Split scan and validation code
