+++ validate.go
@@ -1,0 +1,523 @@
+package jscan
+
+import (
+	"github.com/romshark/jscan/internal/jsonnum"
+	"github.com/romshark/jscan/internal/strfind"
+)
+
+// Valid returns true if s is a valid JSON value.
+//
+// Consider using a reusable Validator instance instead
+// to improve performance when dealing with many inputs.
+func Valid[S ~string | ~[]byte](s S) bool {
+	return !Validate(s).IsErr()
+}
+
+// ValidateOne scans a JSON value from s and returns an error if it's invalid,
+// otherwise returns s with the scanned value cut.
+//
+// Consider using a reusable Validator instance instead
+// to improve performance when dealing with many inputs.
+func ValidateOne[S ~string | ~[]byte](s S) (trailing S, err Error[S]) {
+	var v *Validator[S]
+	switch any(s).(type) {
+	case string:
+		x := validatorPoolString.Get()
+		defer validatorPoolString.Put(x)
+		v = x.(*Validator[S])
+	case []byte:
+		x := validatorPoolBytes.Get()
+		defer validatorPoolBytes.Put(x)
+		v = x.(*Validator[S])
+	}
+	v.stack = v.stack[:0]
+
+	return validate(v.stack, s)
+}
+
+// Validate returns an error if s is invalid JSON.
+//
+// Consider using a reusable Validator instance instead
+// to improve performance when dealing with many inputs.
+func Validate[S ~string | ~[]byte](s S) Error[S] {
+	var v *Validator[S]
+	switch any(s).(type) {
+	case string:
+		x := validatorPoolString.Get()
+		defer validatorPoolString.Put(x)
+		v = x.(*Validator[S])
+	case []byte:
+		x := validatorPoolBytes.Get()
+		defer validatorPoolBytes.Put(x)
+		v = x.(*Validator[S])
+	}
+	v.stack = v.stack[:0]
+
+	t, err := validate(v.stack, s)
+	if err.IsErr() {
+		return err
+	}
+	var illegalChar bool
+	t, illegalChar = strfind.EndOfWhitespaceSeq(t)
+	if illegalChar {
+		return getError(ErrorCodeIllegalControlChar, s, t)
+	}
+	if len(t) > 0 {
+		return getError(ErrorCodeUnexpectedToken, s, t)
+	}
+	return Error[S]{}
+}
+
+// NewValidator creates a new reusable validator instance.
+func NewValidator[S ~string | ~[]byte](preallocStack int) *Validator[S] {
+	return &Validator[S]{
+		stack: make([]stackNodeType, 0, preallocStack),
+	}
+}
+
+// Validator is a reusable validator instance.
+// The validator is more efficient than the parser at JSON validation.
+// A validator instance can be more efficient than global Valid, Validate and ValidateOne
+// functions due to potential stack allocation avoidance.
+type Validator[S ~string | ~[]byte] struct{ stack []stackNodeType }
+
+// Valid returns true if s is a valid JSON value.
+func (v *Validator[S]) Valid(s S) bool {
+	return !v.Validate(s).IsErr()
+}
+
+// ValidateOne scans a JSON value from s and returns an error if it's invalid,
+// otherwise returns s with the scanned value cut.
+func (v *Validator[S]) ValidateOne(s S) (trailing S, err Error[S]) {
+	return validate(v.stack, s)
+}
+
+// Validate returns an error if s is invalid JSON.
+func (v *Validator[S]) Validate(s S) Error[S] {
+	t, err := validate(v.stack, s)
+	if err.IsErr() {
+		return err
+	}
+	var illegalChar bool
+	t, illegalChar = strfind.EndOfWhitespaceSeq(t)
+	if illegalChar {
+		return getError(ErrorCodeIllegalControlChar, s, t)
+	}
+	if len(t) > 0 {
+		return getError(ErrorCodeUnexpectedToken, s, t)
+	}
+	return Error[S]{}
+}
+
+// validate returns the remainder of i.src and an error if any is encountered.
+func validate[S ~string | ~[]byte](st []stackNodeType, s S) (S, Error[S]) {
+	var (
+		rollback S // Used as fallback for error report
+		src      = s
+		top      stackNodeType
+		b        bool
+	)
+
+	stPop := func() {
+		if len(st) < 1 {
+			return
+		}
+		st = st[:len(st)-1]
+	}
+	stTop := func() {
+		if len(st) < 1 {
+			top = 0
+			return
+		}
+		top = st[len(st)-1]
+	}
+	stPush := func(t stackNodeType) {
+		st = append(st, t)
+	}
+
+VALUE:
+	if len(s) < 1 {
+		return s, getError(ErrorCodeUnexpectedEOF, src, s)
+	}
+	if s[0] <= ' ' {
+		switch s[0] {
+		case ' ', '\t', '\r', '\n':
+			s, b = strfind.EndOfWhitespaceSeq(s)
+			if b {
+				return s, getError(ErrorCodeIllegalControlChar, src, s)
+			}
+		}
+		if len(s) < 1 {
+			return s, getError(ErrorCodeUnexpectedEOF, src, s)
+		}
+	}
+	switch s[0] {
+	case '{': // Object
+		goto VALUE_OBJECT
+	case '[': // Array
+		goto VALUE_ARRAY
+	case '-', '0', '1', '2', '3', '4', '5', '6', '7', '8', '9':
+		goto VALUE_NUMBER
+	case '"': // String
+		goto VALUE_STRING
+	case 'n': // Null
+		goto VALUE_NULL
+	case 'f': // False
+		goto VALUE_FALSE
+	case 't': // True
+		goto VALUE_TRUE
+	}
+	if s[0] < 0x20 {
+		return s, getError(ErrorCodeIllegalControlChar, src, s)
+	}
+	return s, getError(ErrorCodeUnexpectedToken, src, s)
+
+VALUE_OBJECT:
+	s = s[1:]
+	if len(s) < 1 {
+		return s, getError(ErrorCodeUnexpectedEOF, src, s)
+	}
+	if s[0] <= ' ' {
+		switch s[0] {
+		case ' ', '\t', '\r', '\n':
+			s, b = strfind.EndOfWhitespaceSeq(s)
+			if b {
+				return s, getError(ErrorCodeIllegalControlChar, src, s)
+			}
+		}
+		if len(s) < 1 {
+			return s, getError(ErrorCodeUnexpectedEOF, src, s)
+		}
+	}
+	if s[0] == '}' {
+		s = s[1:]
+		goto AFTER_VALUE
+	}
+	stPush(stackNodeTypeObject)
+	goto OBJ_KEY
+
+VALUE_ARRAY:
+	stPush(stackNodeTypeArray)
+	s = s[1:]
+	goto VALUE_OR_ARR_TERM
+
+VALUE_NUMBER:
+	{
+		rollback = s
+		if s, b = jsonnum.ReadNumber(s); b {
+			return s, getError(ErrorCodeMalformedNumber, src, rollback)
+		}
+	}
+	goto AFTER_VALUE
+
+VALUE_STRING:
+	s, b = s[1:], false
+	goto STRING
+
+VALUE_NULL:
+	if len(s) < 4 || string(s[:4]) != "null" {
+		return s, getError(ErrorCodeUnexpectedToken, src, s)
+	}
+	s = s[len("null"):]
+	goto AFTER_VALUE
+
+VALUE_FALSE:
+	if len(s) < 5 || string(s[:5]) != "false" {
+		return s, getError(ErrorCodeUnexpectedToken, src, s)
+	}
+	s = s[len("false"):]
+	goto AFTER_VALUE
+
+VALUE_TRUE:
+	if s := s; len(s) < 4 || string(s[:4]) != "true" {
+		return s, getError(ErrorCodeUnexpectedToken, src, s)
+	}
+	s = s[len("true"):]
+	goto AFTER_VALUE
+
+OBJ_KEY:
+	if len(s) < 1 {
+		return s, getError(ErrorCodeUnexpectedEOF, src, s)
+	}
+	if s[0] <= ' ' {
+		switch s[0] {
+		case ' ', '\t', '\r', '\n':
+			s, b = strfind.EndOfWhitespaceSeq(s)
+			if b {
+				return s, getError(ErrorCodeIllegalControlChar, src, s)
+			}
+		}
+		if len(s) < 1 {
+			return s, getError(ErrorCodeUnexpectedEOF, src, s)
+		}
+	}
+	if s[0] != '"' {
+		if s[0] < 0x20 {
+			return s, getError(ErrorCodeIllegalControlChar, src, s)
+		}
+		return s, getError(ErrorCodeUnexpectedToken, src, s)
+	}
+
+	s = s[1:]
+
+	b = true
+	goto STRING
+AFTER_OBJ_KEY_STRING:
+	if len(s) < 1 {
+		return s, getError(ErrorCodeUnexpectedEOF, src, s)
+	}
+	if s[0] <= ' ' {
+		switch s[0] {
+		case ' ', '\t', '\r', '\n':
+			s, b = strfind.EndOfWhitespaceSeq(s)
+			if b {
+				return s, getError(ErrorCodeIllegalControlChar, src, s)
+			}
+		}
+		if len(s) < 1 {
+			return s, getError(ErrorCodeUnexpectedEOF, src, s)
+		}
+	}
+	if s[0] != ':' {
+		if s[0] < 0x20 {
+			return s, getError(ErrorCodeIllegalControlChar, src, s)
+		}
+		return s, getError(ErrorCodeUnexpectedToken, src, s)
+	}
+	s = s[1:]
+	goto VALUE
+
+VALUE_OR_OBJ_TERM:
+	if len(s) < 1 {
+		return s, getError(ErrorCodeUnexpectedEOF, src, s)
+	}
+	if s[0] <= ' ' {
+		switch s[0] {
+		case ' ', '\t', '\r', '\n':
+			s, b = strfind.EndOfWhitespaceSeq(s)
+			if b {
+				return s, getError(ErrorCodeIllegalControlChar, src, s)
+			}
+			goto VALUE_OR_OBJ_TERM
+		}
+		if len(s) < 1 {
+			return s, getError(ErrorCodeUnexpectedEOF, src, s)
+		}
+	}
+	switch s[0] {
+	case '}':
+		s = s[1:]
+		stPop()
+		goto AFTER_VALUE
+	case '{':
+		goto VALUE_OBJECT
+	case '[':
+		goto VALUE_ARRAY
+	case '"':
+		goto VALUE_STRING
+	case 't':
+		goto VALUE_TRUE
+	case 'f':
+		goto VALUE_FALSE
+	case 'n':
+		goto VALUE_NULL
+	case '-', '0', '1', '2', '3', '4', '5', '6', '7', '8', '9':
+		goto VALUE_NUMBER
+	}
+	if s[0] < 0x20 {
+		return s, getError(ErrorCodeIllegalControlChar, src, s)
+	}
+	return s, getError(ErrorCodeUnexpectedToken, src, s)
+
+VALUE_OR_ARR_TERM:
+	if len(s) < 1 {
+		return s, getError(ErrorCodeUnexpectedEOF, src, s)
+	}
+	if s[0] <= ' ' {
+		switch s[0] {
+		case ' ', '\t', '\r', '\n':
+			s, b = strfind.EndOfWhitespaceSeq(s)
+			if b {
+				return s, getError(ErrorCodeIllegalControlChar, src, s)
+			}
+		}
+		if len(s) < 1 {
+			return s, getError(ErrorCodeUnexpectedEOF, src, s)
+		}
+	}
+	switch s[0] {
+	case ']':
+		s = s[1:]
+		stPop()
+		goto AFTER_VALUE
+	case '{':
+		goto VALUE_OBJECT
+	case '[':
+		goto VALUE_ARRAY
+	case '"':
+		goto VALUE_STRING
+	case 't':
+		goto VALUE_TRUE
+	case 'f':
+		goto VALUE_FALSE
+	case 'n':
+		goto VALUE_NULL
+	case '-', '0', '1', '2', '3', '4', '5', '6', '7', '8', '9':
+		goto VALUE_NUMBER
+	}
+	if s[0] < 0x20 {
+		return s, getError(ErrorCodeIllegalControlChar, src, s)
+	}
+	return s, getError(ErrorCodeUnexpectedToken, src, s)
+
+AFTER_VALUE:
+	stTop()
+	if top == 0 {
+		return s, Error[S]{}
+	}
+	if len(s) < 1 {
+		return s, getError(ErrorCodeUnexpectedEOF, src, s)
+	}
+	if s[0] <= ' ' {
+		switch s[0] {
+		case ' ', '\t', '\r', '\n':
+			s, b = strfind.EndOfWhitespaceSeq(s)
+			if b {
+				return s, getError(ErrorCodeIllegalControlChar, src, s)
+			}
+		}
+		if len(s) < 1 {
+			return s, getError(ErrorCodeUnexpectedEOF, src, s)
+		}
+	}
+	switch s[0] {
+	case ',':
+		s = s[1:]
+		if top == stackNodeTypeArray {
+			goto VALUE
+		}
+		goto OBJ_KEY
+	case '}':
+		if top != stackNodeTypeObject {
+			return s, getError(ErrorCodeUnexpectedToken, src, s)
+		}
+		s = s[1:]
+		stPop()
+		goto AFTER_VALUE
+	case ']':
+		if top != stackNodeTypeArray {
+			return s, getError(ErrorCodeUnexpectedToken, src, s)
+		}
+		s = s[1:]
+		stPop()
+		goto AFTER_VALUE
+	}
+	return s, getError(ErrorCodeUnexpectedToken, src, s)
+
+STRING:
+	for {
+		for ; len(s) > 15; s = s[16:] {
+			if lutStr[s[0]] != 0 {
+				goto CHECK_STRING_CHARACTER
+			}
+			if lutStr[s[1]] != 0 {
+				s = s[1:]
+				goto CHECK_STRING_CHARACTER
+			}
+			if lutStr[s[2]] != 0 {
+				s = s[2:]
+				goto CHECK_STRING_CHARACTER
+			}
+			if lutStr[s[3]] != 0 {
+				s = s[3:]
+				goto CHECK_STRING_CHARACTER
+			}
+			if lutStr[s[4]] != 0 {
+				s = s[4:]
+				goto CHECK_STRING_CHARACTER
+			}
+			if lutStr[s[5]] != 0 {
+				s = s[5:]
+				goto CHECK_STRING_CHARACTER
+			}
+			if lutStr[s[6]] != 0 {
+				s = s[6:]
+				goto CHECK_STRING_CHARACTER
+			}
+			if lutStr[s[7]] != 0 {
+				s = s[7:]
+				goto CHECK_STRING_CHARACTER
+			}
+			if lutStr[s[8]] != 0 {
+				s = s[8:]
+				goto CHECK_STRING_CHARACTER
+			}
+			if lutStr[s[9]] != 0 {
+				s = s[9:]
+				goto CHECK_STRING_CHARACTER
+			}
+			if lutStr[s[10]] != 0 {
+				s = s[10:]
+				goto CHECK_STRING_CHARACTER
+			}
+			if lutStr[s[11]] != 0 {
+				s = s[11:]
+				goto CHECK_STRING_CHARACTER
+			}
+			if lutStr[s[12]] != 0 {
+				s = s[12:]
+				goto CHECK_STRING_CHARACTER
+			}
+			if lutStr[s[13]] != 0 {
+				s = s[13:]
+				goto CHECK_STRING_CHARACTER
+			}
+			if lutStr[s[14]] != 0 {
+				s = s[14:]
+				goto CHECK_STRING_CHARACTER
+			}
+			if lutStr[s[15]] != 0 {
+				s = s[15:]
+				goto CHECK_STRING_CHARACTER
+			}
+			continue
+		}
+
+	CHECK_STRING_CHARACTER:
+		if len(s) < 1 {
+			return s, getError(ErrorCodeUnexpectedEOF, src, s)
+		}
+		switch s[0] {
+		case '\\':
+			if len(s) < 2 {
+				return s, getError(ErrorCodeUnexpectedEOF, src, s)
+			}
+			if lutEscape[s[1]] == 1 {
+				s = s[2:]
+				continue
+			}
+			if s[1] != 'u' {
+				return s, getError(ErrorCodeInvalidEscapeSeq, src, s)
+			}
+			if len(s) < 6 ||
+				lutSX[s[5]] != 2 ||
+				lutSX[s[4]] != 2 ||
+				lutSX[s[3]] != 2 ||
+				lutSX[s[2]] != 2 {
+				return s, getError(ErrorCodeInvalidEscapeSeq, src, s)
+			}
+			s = s[5:]
+		case '"':
+			s = s[1:]
+			if b {
+				goto AFTER_OBJ_KEY_STRING
+			}
+			goto AFTER_VALUE
+		default:
+			if s[0] < 0x20 {
+				return s, getError(ErrorCodeIllegalControlChar, src, s)
+			}
+			s = s[1:]
+		}
+	}
+}
